+++ tests/bfabric_app_runner/specs/test_inputs_spec.py
@@ -1,9 +1,10 @@
 import pytest
 import yaml
 
-from bfabric_app_runner.specs.inputs_spec import InputsSpec
+from bfabric.entities import Resource
 from bfabric_app_runner.specs.inputs.bfabric_dataset_spec import BfabricDatasetSpec
 from bfabric_app_runner.specs.inputs.bfabric_resource_spec import BfabricResourceSpec
+from bfabric_app_runner.specs.inputs_spec import InputsSpec
 
 
 @pytest.fixture()
@@ -43,3 +44,24 @@ def test_serialize(parsed, serialized):
 
 def test_parse(parsed, serialized):
     assert InputsSpec.model_validate(yaml.safe_load(serialized)) == parsed
+
+
+def test_resolve_filename_with_filename(mocker):
+    """Test when filename is provided"""
+    client = mocker.Mock(name="client")
+    spec = BfabricResourceSpec(id=1, filename="test.txt", check_checksum=True)
+
+    result = spec.resolve_filename(client)
+    assert result == "test.txt"
+
+
+def test_resolve_filename_without_filename(mocker):
+    """Test when filename is not provided"""
+    client = mocker.MagicMock(name="client")
+    mock_resource_find = mocker.patch.object(Resource, "find", return_value={"relativepath": "/path/to/data.csv"})
+
+    spec = BfabricResourceSpec(id=1, check_checksum=True)
+
+    result = spec.resolve_filename(client)
+    assert result == "data.csv"
+    mock_resource_find.assert_called_once_with(id=1, client=client)