+++ bfabric_app_runner/docs/changelog.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - App Definition now supports omitting the collect step.
 
+### Changed
+
+- `BfabricResourceSpec` defaults to file basename vs resource name, if no name is specified.
+
 ## \[0.0.17\] - 2025-02-19
 
 ### Fixed